run integration tests with maven failsafe plugin, apply spotless to tests
