+++ src/test/resources/build-info.properties
@@ -1,0 +1,1 @@
+build.version=${project.version}
